refactor: extraer algoritmo de simplificacion de deudas a SaldoCalculator
Se mueve el calculo de saldos netos y su simplificacion en transferencias
minimas desde BalanceService (metodos privados) a un componente reutilizable
SaldoCalculator, para que el motor de liquidacion de periodos pueda usar el
mismo algoritmo sin duplicarlo. Tambien se corrige un bug preexistente en
BalanceService.registrarPago que ignoraba el monto parcial informado y usaba
siempre el saldo total de la deuda, y se agrega Gasto.agregarParticipante
(faltante tras un refactor anterior) para mantener la relacion bidireccional
con GastoParticipante.

+++ src/test/java/com/example/fairshareapp/controller/BalanceControllerTest.java
@@ -1,5 +1,7 @@
 package com.example.fairshareapp.controller;
 
+import java.math.BigDecimal;
+
 import com.example.fairshareapp.exception.GlobalExceptionHandler;
 import com.example.fairshareapp.exception.RecursoNoEncontradoException;
 import com.example.fairshareapp.exception.ReglaInvalidaException;
@@ -63,7 +65,7 @@ class BalanceControllerTest {
                 .deudorNombre("Maria Gomez")
                 .acreedorId(1L)
                 .acreedorNombre("Juan Perez")
-                .monto(500.0)
+                .monto(BigDecimal.valueOf(500.0))
                 .estado(EstadoDeuda.PENDIENTE)
                 .build();
 
@@ -101,7 +103,7 @@ class BalanceControllerTest {
         DeudaDetalleDTO deuda = DeudaDetalleDTO.builder()
                 .id(1L)
                 .estado(EstadoDeuda.SALDADO)
-                .monto(0.0)
+                .monto(BigDecimal.valueOf(0.0))
                 .build();
 
         when(balanceService.registrarPago(eq(1L), any())).thenReturn(deuda);
